Compress packet serializers (#2289)
* WIP packet template serializers

* More packets

* More packets

* More packets

* No more packets

* rename to buffer

* less generic mess

* review stuff

* Fix ClientResourcePackStatusPacket enum

* Update UpdateEnabledFeaturesPacket

+++ src/main/java/net/minestom/server/entity/Player.java
@@ -525,7 +525,7 @@ public class Player extends LivingEntity implements CommandSender, Localizable, 
         refreshHealth();
 
         sendPacket(new RespawnPacket(dimensionTypeId, instance.getDimensionName(), 0, gameMode, gameMode,
-                false, levelFlat, deathLocation, portalCooldown, RespawnPacket.COPY_ALL));
+                false, levelFlat, deathLocation, portalCooldown, (byte) RespawnPacket.COPY_ALL));
         refreshClientStateAfterRespawn();
 
         PlayerRespawnEvent respawnEvent = new PlayerRespawnEvent(this);
@@ -1208,7 +1208,7 @@ public class Player extends LivingEntity implements CommandSender, Localizable, 
         final PlayerInfoUpdatePacket addPlayerPacket = getAddPlayerToList();
 
         RespawnPacket respawnPacket = new RespawnPacket(dimensionTypeId, instance.getDimensionName(),
-                0, gameMode, gameMode, false, levelFlat, deathLocation, portalCooldown, RespawnPacket.COPY_ALL);
+                0, gameMode, gameMode, false, levelFlat, deathLocation, portalCooldown, (byte) RespawnPacket.COPY_ALL);
 
         sendPacket(removePlayerPacket);
         sendPacket(destroyEntitiesPacket);
@@ -1658,7 +1658,7 @@ public class Player extends LivingEntity implements CommandSender, Localizable, 
         this.dimensionTypeId = DIMENSION_TYPE_REGISTRY.getId(dimensionType);
         sendPacket(new RespawnPacket(dimensionTypeId, dimensionName,
                 0, gameMode, gameMode, false, levelFlat,
-                deathLocation, portalCooldown, RespawnPacket.COPY_ALL));
+                deathLocation, portalCooldown, (byte) RespawnPacket.COPY_ALL));
         refreshClientStateAfterRespawn();
     }
